added js for products
Added js for side-bar and for links cart and favorites (for clicks on the document and checks if the clicked element has the class add-to-cart or add-to-favorites. If it does, it retrieves the product ID from the data-product-id attribute of the closest .box element and calls the appropriate function (addToCart or addToFavorites)).